requirements.txt
#common
opencv-python==4.1.1.26
easydict

# for server
absl-py
# tensorflow
tensorflow-gpu==2.3.0
matplotlib
numpy<1.19.0
pandas
scipy==1.4.1
protobuf==3.9.2

# for api
fastapi
uvicorn
jinja2
python-multipart